+++ src/components/Field/Field.tsx
@@ -1,9 +1,14 @@
 import React from 'react';
-// import Button from '@material-ui/core/Button';
 import {connect} from 'react-redux';
 import Cell from '../Cell/Cell';
 
-class Field extends React.Component {
+interface IFoundation {
+    foundation:number,
+    onSetFoundation:any,
+    onGenerateField:any,
+}
+
+class Field extends React.Component<IFoundation,{}> {
 
     generateField(){
         let field = [];
@@ -12,7 +17,7 @@ class Field extends React.Component {
             for(var j = 0; j<8; j++){
                 row.push(<Cell key={`${i},${j}`}/>)
             }
-        field.push(<div className="row">{row}</div>)
+        field.push(<div className="row" key={i}>{row}</div>)
         }
 
         return (
@@ -23,6 +28,7 @@ class Field extends React.Component {
     }
 
     render(){
+        console.log('@@@@@@@@@@ ', this.props.foundation)
         return (
             <div>{this.generateField()}</div>
         );
@@ -33,7 +39,7 @@ class Field extends React.Component {
 
 function mapStateToProps(state:any){
     return {
-        foundation: state.foundation
+        foundation: state.foundationStore.foundation
     }
 }
