+++ src/interfaces/Foundation.ts
@@ -1,0 +1,5 @@
+export interface IFoundation {
+    foundation:number,
+    onSetFoundation:any,
+    onGenerateField:any,
+}
