+++ src/redux/reducers/RootReducer.tsx
@@ -1,19 +1,7 @@
-interface IAppStore {
-  foundation: number;
-}
+import {combineReducers} from 'redux';
+import foundationStore from './FoundationReducer';
+import turnStore from './TurnReducer';
 
-const initialState:IAppStore = {
-    foundation: 2,
-}
-
-export default function foundationStore(state = initialState, action:any):IAppStore{
-
-    switch(action.type){
-        case 'SET_FOUNDATION':
-            return {
-              foundation: action.foundation,
-            }
-        default:
-          return state
-    }
-}
+export default combineReducers({
+  foundationStore, turnStore
+})
